moved test in sepparate file, replaced naked strings
