+++ validate.py
@@ -5,7 +5,7 @@ import os
 import unittest
 import sys
 
-import v1.policy_manager.validatepolicy as validatepolicy
+import v1.policy_manager.test_validatepolicy as validatepolicy
 import v1.dynamic_lists.validate_dynamic_lists as validate_dynamic_lists
 import v1.captiveportal.validate_captiveportal as validate_captiveportal
 import v1.system.validate_system_schema as validate_system_schema
